Add files via upload
In this innovative project, I dive into the realm of machine learning to develop a predictive model capable of estimating car prices with remarkable accuracy. By leveraging advanced machine learning algorithms and a comprehensive dataset, my goal is to create a model that can provide precise and reliable price predictions for various car models. This project demonstrates the powerful applications of machine learning in the automotive industry and beyond.